pyproject.toml
[project]
name = "companion-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-sqlite>=3.0.2",
    "python-dotenv>=1.2.1",
    "pywebpush>=2.2.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
